Now resizable, print values on axis, dashed line from axis to vertex

+++ src/example/Example.java
@@ -11,8 +11,8 @@ import java.awt.*;
 public class Example extends JFrame {
 
     public Example() {
-        long[] x = {0, 1, 2, 3, 4, 5, 6, 7, 8};
-        long[] y = {55, 348, 246, 266, 116, 158, 299, 313, 90};
+        double[] x = {0, 1, 2, 3, 4, 5, 6, 7, 8};
+        double[] y = {55, 348, 246, 266, 116, 158, 299, 313, 90};
 
         JLineGraph chart = new JLineGraph(x, y, "Example", 800, 600);
         chart.setLineColor(Color.green);
